pyproject.toml
[tool.poetry]
name = "django-samurai"
version = "0.1.0"
description = "Integrated views and templates, type-safe templates, and file-based routing for HTMX developers with deadlines."
authors = ["Christian Tanul <contact@tanulchristian.dev>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django-ninja = "^1.1.0"
django = "^5.0.6"

[tool.poetry.group.dev.dependencies]
ruff = "^0.4.8"
pre-commit = "^3.7.1"
django-pytest = "^0.2.0"
pytest-cov = "^5.0.0"
pyupgrade = "^3.15.2"
selenium = "^4.21.0"
pytest-playwright = "^0.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
